+++ package/src/response/json/functions/mod.rs
@@ -1,3 +1,0 @@
-pub mod create;
-pub mod failure;
-pub mod success;