+++ package/src/response/json/mod.rs
@@ -1,14 +1,13 @@
 pub(crate) mod create;
 pub(crate) mod error;
-pub(crate) mod functions;
 pub(crate) mod response;
 
 pub use crate::response::json::error::{JsonResponseError, ResponseError};
 
 pub use crate::response::json::response::JsonResponse;
 
-pub use crate::response::json::functions::success::JsonSuccessResponseFunctions;
+pub use crate::response::json::create::success::CreateSuccessJsonResponse;
 
-pub use crate::response::json::functions::failure::JsonFailureResponseFunctions;
+pub use crate::response::json::create::failure::CreateFailureJsonResponse;
 
 pub use crate::response::json::create::CreateJsonResponse;